requirements.txt
altgraph==0.17.4
audioop-lts==0.2.1
audioread==3.0.1
certifi==2025.4.26
cffi
charset-normalizer==3.4.2
colorednoise==2.2.0
contourpy
cycler==0.12.1
decorator==5.2.1
et_xmlfile==2.0.0
fonttools==4.58.0
idna==3.10
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
macholib==1.16.3
matplotlib
msgpack==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
pydeps==3.0.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.3
PyQt5-Qt5
PyQt5_sip
PyQtWebEngine
PyQtWebEngine-Qt5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
stdlib-list==0.11.1
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0